requirements.txt
streamlit
requests
PyPDF2
sentence-transformers
faiss-cpu
google.generativeai

